requirements.txt
# ============================================================
#  APLIKASI PEMINJAMAN ALAT — Dependencies
#  Install: pip install -r requirements.txt
# ============================================================

# Core
Django==4.2.10
djangorestframework==3.14.0

# Database
mysqlclient==2.2.4

# Authentication JWT
djangorestframework-simplejwt==5.3.1

# Environment variables
django-environ==0.11.2

# Filtering & Search
django-filter==23.5

# API Documentation (Swagger)
drf-spectacular==0.27.1

# QR Code generator
qrcode[pil]==7.4.2
Pillow==11.1.0

# Rate Limiting
django-ratelimit==4.1.0

# Password hashing (bcrypt backend)
bcrypt==4.1.2

# CORS (untuk frontend nanti)
django-cors-headers==4.3.1

# Dev tools
ipython==8.22.1
